[FEAT] #21 - 호출탭 : 호출하기 서버 연결

+++ app/src/main/java/com/example/cpr2u_android/data/model/request/education/RequestCall.kt
@@ -7,7 +7,7 @@ data class RequestCall(
     @SerializedName("full_address")
     val fullAddress: String,
     @SerializedName("latitude")
-    val latitude: Int,
+    val latitude: Double,
     @SerializedName("longitude")
-    val longitude: Int
+    val longitude: Double
 )
